하나모아/FastAPI/하나모아-ml/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
numpy>=1.24.0,<2.0.0
pandas>=2.0.0,<3.0.0
scikit-learn>=1.3.0,<2.0.0
mysql-connector-python>=8.1.0,<9.0.0
python-multipart==0.0.6
python-dotenv==1.0.0
joblib>=1.3.0
